Add back in explicit (double) cast for smartass.rol
The value is of type MuMech.EditableDouble, which String.Format doesn't handle (it instead displays "MuMech.EditableDouble").

+++ MechJebRPM/MechJebRPM.cs
@@ -185,7 +185,7 @@ namespace MechJebRPM
 			topMenu.Add(new TextMenu.Item(MechJebModuleSmartASS.ModeTexts[(int)MechJebModuleSmartASS.Mode.ORBITAL], OrbitalMenu));
 			targetMenuItem = new TextMenu.Item(MechJebModuleSmartASS.ModeTexts[(int)MechJebModuleSmartASS.Mode.TARGET], TargetMenu);
 			topMenu.Add(targetMenuItem);
-			forceRollMenuItem = new TextMenu.Item(String.Format("Force Roll: {0:f0}", activeSmartass.rol), ToggleForceRoll);
+			forceRollMenuItem = new TextMenu.Item(String.Format("Force Roll: {0:f0}", (double)activeSmartass.rol), ToggleForceRoll);
 			topMenu.Add(forceRollMenuItem);
 			executeNodeItem = new TextMenu.Item("Execute Next Node", ExecuteNode);
 			topMenu.Add(executeNodeItem);
@@ -203,7 +203,7 @@ namespace MechJebRPM
 
 			targetMenuItem.isDisabled = (FlightGlobals.fetch.VesselTarget == null);
 			nodeMenuItem.isDisabled = (vessel.patchedConicSolver.maneuverNodes.Count == 0);
-			forceRollMenuItem.labelText = String.Format("Force Roll - {0:f0}", activeSmartass.rol);
+			forceRollMenuItem.labelText = String.Format("Force Roll - {0:f0}", (double)activeSmartass.rol);
 
 			MechJebModuleManeuverPlanner mp = null;
 			if (activeJeb != null) {
